requirements.txt
# ─────────────────────────────────────────────────────
# Core Web Framework and API Utilities
# ─────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests>=2.32.3
aiohttp>=3.10.5
httpx>=0.27.2
websocket-client>=1.6.0
